+++ PDKitTests/Dates/DateExtensionTests.swift
@@ -18,11 +18,9 @@ class DateExtensionTests: XCTestCase {
     // For example, the tests may assume it can add seconds to the current time and have it still be today.
 
     func testDayValue() {
-        let defaultDate = DateFactory.createDefaultDate()  // DayValue is 31
-        let twoDaysWorthOfHours = 48
-        let date = DateFactory.createDate(byAddingHours: twoDaysWorthOfHours, to: defaultDate)
-        let actual = date!.dayValue()
-        let expected = 2  //  day 31 + 1 day = day 1, + 1 day = day 2
+        let defaultDate = DateFactory.createDefaultDate()
+        let actual = defaultDate.dayValue()
+        let expected = 1
         XCTAssertEqual(expected, actual)
     }